MyShop/requirements.txt
pillow==10.3.0
pip==23.2.1
wheel==0.41.2
tornado==6.4
amqp==5.2.0
vine==5.1.0
pytz==2024.1
Django==5.0.6
kombu==5.3.7
billiard==4.2.0
celery==5.4.0
click==8.1.7
python-dateutil==2.9.0.post0
asgiref==3.8.1
sqlparse==0.5.0
colorama==0.4.6
humanize==4.9.0
prometheus_client==0.20.0
flower==2.0.1
wcwidth==0.2.13
six==1.16.0
setuptools==68.2.0
psycopg2==2.9.9
sorl-thumbnail==12.10.0
djangorestframework==3.15.2
